requirements.txt
# Core dependencies
langgraph>=0.0.10
discord.py>=2.3.2
qdrant-client>=1.7.0
langchain-groq>=0.0.3
langchain-core>=0.1.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0  # For better performance with fuzzywuzzy

# Embedding and ML dependencies
huggingface-hub>=0.20.1
transformers>=4.36.0
sentence-transformers>=2.2.2  # For multilingual-e5-small model

# Utility dependencies
python-dotenv>=1.0.0  # For environment variables management
requests>=2.31.0      # For HTTP requests
asyncio>=3.4.3       # For async operations
uuid>=1.30           # For generating unique identifiers

# Optional but recommended
typing-extensions>=4.8.0  # For better type hints